WIP: converted to module, stricter linting one test passing
